+++ src/Services/HookService.php
@@ -143,7 +143,12 @@ class HookService
 
     private function shouldUseHttpProtocolForHost(string $host)
     {
-        if (in_array($host, ['localhost', '127.0.0.1', '::1'], true)) {
+        if (
+            in_array($host, ['localhost', '127.0.0.1', '::1'], true)
+            || str_ends_with($host, '.localhost')
+            || str_ends_with($host, '.test')
+            || str_ends_with($host, '.local')
+        ) {
             return true;
         }